x86/mm/pat: Make failure message of memtype_free() a warning
memtype_free() can only fail if its parameters are invalid, and its callers
should already ensure that to not be the case. If it does fail, there must
be a kernel bug somewhere. However, in that case, there is only a
pr_info(), which can easily be overlooked.

Change it to pr_warn(), so that it is more noticeable.

Signed-off-by: Nam Cao <namcao@linutronix.de>